pyproject.toml
[project]
name = "context-kubernetes"
version = "0.1.0"
description = "Context Kubernetes: Declarative Orchestration of Enterprise Knowledge for Agentic AI Systems"
authors = [{name = "Charafeddine Mouzouni", email = "charafeddine@cohorte.co"}]
requires-python = ">=3.11"
license = {text = "Apache-2.0"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "redis>=5.2.0",
    "pyjwt[crypto]>=2.10.0",
    "httpx>=0.28.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "pyyaml>=6.0.0",
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
llm = [
    "openai>=1.55.0",
    "anthropic>=0.40.0",
]
gmail = [
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.155.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
